视屏首页

+++ app/src/main/java/com/qd/welfare/entity/VideoInfo.java
@@ -28,7 +28,7 @@ public class VideoInfo implements Serializable {
     private String thumb;
     private String thumb_shu;
     private int type;
-    private int star;
+    private float star;
     private int play_times;
     private String url;
     private long duration;
@@ -74,11 +74,11 @@ public class VideoInfo implements Serializable {
         this.type = type;
     }
 
-    public int getStar() {
+    public float getStar() {
         return star;
     }
 
-    public void setStar(int star) {
+    public void setStar(float star) {
         this.star = star;
     }
